make bootloader more unlikely

+++ boost_g474/Src/main.c
@@ -24,7 +24,7 @@
 /* Private includes ----------------------------------------------------------*/
 /* USER CODE BEGIN Includes */
 #include "../../motorlib/system.h"
-uint8_t go_to_bootloader = 0;
+uint32_t go_to_bootloader = 0;
 /* USER CODE END Includes */
 
 /* Private typedef -----------------------------------------------------------*/
